Prediction and simulation method for 'mobility.model' class

This function uses a fitted mobility.model object to simulate a connectivity matrix based on estimated parameters.

predict(object, newdata, nsim, seed, ...)

Arguments

object
a mobility.model object containing a fitted mobility model and its associated data

newdata
a list containing new data to used in model prediction. If NULL (the default) the function will simulate the model with the data used for fitting.

D

a named matrix of distances between all \(ij\) location pairs

N

a named vector of population sizes for all locations (either N or both N_orig and N_dest must be supplied)

N_orig

a named vector of population sizes for each origin

N_dest

a named vector of population sizes for each destination

nsim
number of simulations (default = 1).

seed
optional integer specifying the call to set.seed prior to model simulation (default = NULL)

...
further arguments passed to or from other methods

Value

a vector, matrix, or array containing predited or simulated mobility values.

Details

When nsim = 1, the prediction matrix is calculated using the mean point estimate of parameter values. If nsim > 1 then returns and array that contains nsim number of simulated replications based on the posterior distributions of each parameter.

See also

Author

John Giles

Examples

mod <- mobility(data=mobility_matrices, model='gravity', type='transport')

#> ::Fitting transport gravity model for 10 origins and 10 destinations::

#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 70
#>    Unobserved stochastic nodes: 32
#>    Total graph size: 417
#> 
#> Initializing model
#> 
#> NOTE: Stopping adaptation
#> 
#> 

predict(object=mod)

#>       destination
#> origin         A         B         C         D         E         F         G
#>      A 5265.4046  443.3968  240.1566  468.3943  575.8860  338.5657  553.0002
#>      B  443.3968 4891.3172  279.4241  388.4593  617.8456  408.6413  501.3102
#>      C  240.1566  279.4241 1672.1655  216.8236  355.1774  226.6721  261.4031
#>      D  468.3943  388.4593  216.8236 4979.1355  515.7756  300.4047  444.9664
#>      E  575.8860  617.8456  355.1774  515.7756 8777.6952  539.6385  631.4157
#>      F  338.5657  408.6413  226.6721  300.4047  539.6385 2993.8916  377.1612
#>      G  553.0002  501.3102  261.4031  444.9664  631.4157  377.1612 5549.8097
#>      H  647.6914  605.1104  312.5601  526.5075  741.1022  447.4023  832.3066
#>      I  302.9328  279.6737  144.8388  245.0724  344.6253  207.5663  401.5971
#>      J  308.3946  279.7355  157.5772  287.4152  396.9271  222.6213  315.9904
#>       destination
#> origin         H         I         J
#>      A  647.6914  302.9328  308.3946
#>      B  605.1104  279.6737  279.7355
#>      C  312.5601  144.8388  157.5772
#>      D  526.5075  245.0724  287.4152
#>      E  741.1022  344.6253  396.9271
#>      F  447.4023  207.5663  222.6213
#>      G  832.3066  401.5971  315.9904
#>      H 7878.7980  614.9266  367.9567
#>      I  614.9266 1698.1252  171.5629
#>      J  367.9567  171.5629 2351.1013
#> attr(,"model")
#> [1] "gravity"
#> attr(,"type")
#> [1] "transport"

n <- 5
ids <- letters[1:n]

# Distance matrix
D <- get_distance_matrix(x=rnorm(n, -100, 2),
                         y=rnorm(n, 20, 1),
                         id=ids)*111.35

# Vector of population sizes
N <- rnbinom(n, size=5, mu=5000)
names(N) <- ids

# Predict mobility model using new data
predict(object=mod, newdata=list(D=D, N=N))

#>       destination
#> origin         a         b         c          d         e
#>      a 2514.2870  283.1106 106.08746  215.30098 123.18338
#>      b  283.1106 2962.5905 125.99328  257.81314 126.08975
#>      c  106.0875  125.9933 501.02888   89.05014  48.45009
#>      d  215.3010  257.8131  89.05014 1759.27000  90.15599
#>      e  123.1834  126.0898  48.45009   90.15599 456.31604
#> attr(,"model")
#> [1] "gravity"
#> attr(,"type")
#> [1] "transport"