Extract model residuals

Generic fundtion that extracts model residuals from a 'mobility.model' object.

residuals(object, type, ...)

Arguments

object
a mobility.model object produced by the mobility function

type
the type of residuals to be returned. Available residual types include: 'deviance' (default), 'pearson', and 'raw'.

...
further arguments passed to or from other methods

Value

a matrix containing model residuals

Details

Residual types are calculated as:

raw

\(y_i - \mu_i\)

pearson

\((y_i - \mu_i)/\sqrt{\mu_i}\)

deviance

\(sign(y_i - \mu_i) * \sqrt(2(log(y_i/\mu_i) - (y_i - \mu_i)))\)

Where, \(y_i\) is the observed data and \(\mu_i\) is the value predicted by the model using the mean of parameter posterior distributions

See also

Author

John Giles

Examples

mod <- mobility(data=mobility_matrices, model='gravity', type='transport', DIC=TRUE)

#> ::Fitting transport gravity model for 10 origins and 10 destinations::

#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 70
#>    Unobserved stochastic nodes: 32
#>    Total graph size: 417
#> 
#> Initializing model
#> 
#> NOTE: Stopping adaptation
#> 
#> 

residuals(mod, type='pearson')

#>       destination
#> origin          A           B          C           D          E          F
#>      A  -9.732584  -3.4381789  -8.269817   0.9058579  -9.829610  -3.726410
#>      B         NA  58.9370234 -13.186432 -18.6946645 -22.040351         NA
#>      C -13.883801          NA         NA -14.3174955 -16.564532 -11.003991
#>      D  69.105205   7.5872000         NA -37.7960654   6.438467   6.034657
#>      E         NA   0.2475687   1.051786 -16.6343181  -1.882608         NA
#>      F -16.552372  -6.6604894  -6.952351          NA -10.229745         NA
#>      G -15.096137 -16.8964345         NA          NA -21.944310 -15.867733
#>      H -19.398651 -20.1679247         NA          NA -24.431484 -18.267999
#>      I -15.394044 -15.3481401 -11.702550 -14.8243964 -17.809959 -13.227203
#>      J  23.210544          NA   3.618435  11.9494995  23.895536         NA
#>       destination
#> origin          G         H          I          J
#>      A  32.318485  24.72746  26.548013  -3.496079
#>      B -19.218903 -20.94031 -14.212011 -16.067636
#>      C -14.498018        NA -10.954726 -11.915701
#>      D  32.474832        NA         NA  26.223967
#>      E  -7.458483        NA  -3.804435         NA
#>      F         NA        NA         NA -12.909869
#>      G  21.536699  13.47316         NA         NA
#>      H   3.594369 -41.98368 239.703657 -17.670405
#>      I  -7.115595        NA         NA         NA
#>      J  26.102911  16.37153         NA         NA
#> attr(,"model")
#> [1] "gravity"
#> attr(,"type")
#> [1] "transport"
#> attr(,"residuals")
#> [1] "pearson"