Simulate probability of leaving origin

This function simulates one stochastic realization of the proportion of individuals that leave origin \(i\). The function takes posterior estimates of the mean (mu) and standard deviation (sigma) for each origin \(i\), and then derives the shape and rate parameters for the Beta distribution. Simulated values are random draws from this Beta distribution.

sim_prob_travel(mu = 0.5, sigma = 0.1, n = 1, id = NULL)

Arguments

mu
scalar or vector giving mean probability of leaving origin (defualt = 0.5)

sigma
scalar or vector giving standard deviation of the probability of leaving origin (default = 0.1)

n
number of simulations (default = 1)

id
optional scalar or vector giving name(s) of origin (default = NULL)

Value

a numeric scalar or vector with values between 0 and 1

See also

Examples

sim_prob_travel(0.5, 0.04)

#> [1] 0.5378499

sim_prob_travel(mu=c(0.8, 0.6),
                sigma=c(0.08, 0.06),
                id=c('A', 'B'))

#>         A         B 
#> 0.5410971 0.6071538 

# Simulate with estimated parameters
n_orig <- 6
n_missing <- 3
orig_id <- LETTERS[1:n_orig]

N <- rpois(n_orig, 100)    # population size of each origin
p <- rbeta(n_orig, 2, 2)   # probability of leaving origin

travel <- setNames(rbinom(n_orig, N, p), orig_id)
total <- setNames(N, orig_id)

miss <- sample(1:n_orig, n_missing) # missing observations
travel[miss] <- total[miss] <- NA

# Estimate probability of travel for each locations (missing locations regress to mean)
prob_trav <- summarize_mobility(
  fit_prob_travel(travel=travel, total=total)
)

#> These missing locations will inherit population mean:

#> A B F

#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 3
#>    Unobserved stochastic nodes: 6
#>    Total graph size: 17
#> 
#> Initializing model
#> 
#> NOTE: Stopping adaptation
#> 
#> 

sim_prob_travel(mu=prob_trav$Mean,
                sigma=prob_trav$SD,
                id=names(travel),
                n=5)

#>        [,1]      [,2]      [,3]      [,4]      [,5]
#> A 0.3542443 0.6718827 0.4426144 0.2023465 0.4474605
#> B 0.2951938 0.4550604 0.4439437 0.6430328 0.3908564
#> C 0.2884857 0.3057066 0.3592183 0.2845741 0.3036692
#> D 0.1555084 0.2374283 0.2475775 0.2159101 0.2555146
#> E 0.5830917 0.6806101 0.5919640 0.5712152 0.6119025
#> F 0.4180785 0.5873098 0.6519638 0.3103818 0.4381072