Prediction and simulation method for 'mobility.model' class

This function uses a fitted mobility.model object to simulate a connectivity matrix based on estimated parameters.

predict(object, newdata, nsim, seed, ...)

Arguments

object
a mobility.model object containing a fitted mobility model and its associated data

newdata
a list containing new data to used in model prediction. If NULL (the default) the function will simulate the model with the data used for fitting.

D

a named matrix of distances between all \(ij\) location pairs

N

a named vector of population sizes for all locations (either N or both N_orig and N_dest must be supplied)

N_orig

a named vector of population sizes for each origin

N_dest

a named vector of population sizes for each destination

nsim
number of simulations (default = 1).

seed
optional integer specifying the call to set.seed prior to model simulation (default = NULL)

...
further arguments passed to or from other methods

Value

a vector, matrix, or array containing predited or simulated mobility values.

Details

When nsim = 1, the prediction matrix is calculated using the mean point estimate of parameter values. If nsim > 1 then returns and array that contains nsim number of simulated replications based on the posterior distributions of each parameter.

See also

Author

John Giles

Examples

mod <- mobility(data=mobility_matrices, model='gravity', type='transport')

#> ::Fitting transport gravity model for 10 origins and 10 destinations::

#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 70
#>    Unobserved stochastic nodes: 32
#>    Total graph size: 421
#> 
#> Initializing model
#> 
#> NOTE: Stopping adaptation
#> 
#> 

predict(object=mod)

#>       destination
#> origin         A         B         C         D         E         F         G
#>      A 5270.8807  444.0193  240.4966  469.0487  576.6973  339.0423  553.7636
#>      B  444.0193 4896.4043  279.8118  389.0107  618.7061  409.2028  502.0069
#>      C  240.4966  279.8118 1673.9046  217.1330  355.6729  226.9853  261.7708
#>      D  469.0487  389.0107  217.1330 4984.3139  516.5085  300.8317  445.5932
#>      E  576.6973  618.7061  355.6729  516.5085 8786.8242  540.3811  632.2990
#>      F  339.0423  409.2028  226.9853  300.8317  540.3811 2997.0053  377.6875
#>      G  553.7636  502.0069  261.7708  445.5932  632.2990  377.6875 5555.5817
#>      H  648.5872  605.9502  312.9996  527.2498  742.1405  448.0269  833.4298
#>      I  303.3514  280.0620  145.0426  245.4179  345.1080  207.8561  402.1367
#>      J  308.8275  280.1306  157.8007  287.8204  397.4844  222.9353  316.4340
#>       destination
#> origin         H         I         J
#>      A  648.5872  303.3514  308.8275
#>      B  605.9502  280.0620  280.1306
#>      C  312.9996  145.0426  157.8007
#>      D  527.2498  245.4179  287.8204
#>      E  742.1405  345.1080  397.4844
#>      F  448.0269  207.8561  222.9353
#>      G  833.4298  402.1367  316.4340
#>      H 7886.9921  615.7298  368.4745
#>      I  615.7298 1699.8913  171.8042
#>      J  368.4745  171.8042 2353.5465
#> attr(,"model")
#> [1] "gravity"
#> attr(,"type")
#> [1] "transport"

n <- 5
ids <- letters[1:n]

# Distance matrix
D <- get_distance_matrix(x=rnorm(n, -100, 2),
                         y=rnorm(n, 20, 1),
                         id=ids)*111.35

# Vector of population sizes
N <- rnbinom(n, size=5, mu=5000)
names(N) <- ids

# Predict mobility model using new data
predict(object=mod, newdata=list(D=D, N=N))

#>       destination
#> origin         a         b         c          d         e
#>      a 2516.9019  283.5008 106.23517  215.59818 123.35128
#>      b  283.5008 2965.6717 126.16701  258.16524 126.26272
#>      c  106.2352  126.1670 501.54996   89.17408  48.51704
#>      d  215.5982  258.1652  89.17408 1761.09968  90.28067
#>      e  123.3513  126.2627  48.51704   90.28067 456.79062
#> attr(,"model")
#> [1] "gravity"
#> attr(,"type")
#> [1] "transport"