Simulate a full mobility model

This function simulates a mobility matrix based on the supplied model parameters. The full mobility model estimates both quantity of travel within a location (along the matrix diagonal) as well as the quantity among all origins and destinations (off-diagonals). For multiple stochastic simulations, the mod argument must be supplied with a mobility model object produced by the fit_mobility function. If mean values are supplied to the named parameter arguments (theta, omega_1, omega_2, gamma, tau), a single point estimate of the mobility matrix will be returned. A null model (where all model parameters = 1) can be simulated by supplying only population sizes (N) and pairwise distances (D).

sim_mobility(
  D,
  N,
  theta = 1,
  omega_1 = 1,
  omega_2 = 1,
  gamma = 1,
  tau = 0.5,
  n = 1,
  mod = NULL,
  counts = FALSE
)

Arguments

D
matrix giving distances among the origins and destinations

N
vector of population sizes

theta
scalar giving the proportionality constant of gravity formula (default = 1)

omega_1
scalar giving exponential scaling of origin population size (default = 1)

omega_2
scalar giving exponential scaling of destination population size (default = 1)

gamma
scalar giving the dispersal kernel paramater (default = 1)

tau
scalar or vector giving the probability of travel outside origin

n
number of simulations (requires argument mod to be supplied with a gravity model object)

mod
a mobility model object produced by the fit_mobility or summarize_mobility functions

counts
logical indicating whether or not to return a count variable by scaling the mobility matrix by origin population size (\(N_i\)) (default = FALSE)

Value

a matrix with values between 0 and 1 (if counts = FALSE) or positive integers (if counts = TRUE). If n > 1 then returns and array with 3 dimensions

See also

Examples

M <- mobility_matrices$M
D <- mobility_matrices$D
N <- mobility_matrices$N

# Simulate null model for given D and N
# assumes 0.5 probability of traveling outside origin and null gravity
# for probability of travel to all destinations
sim_mobility(D, N)

#>       destination
#> origin          A          B           C           D           E           F
#>      A 0.50000000 0.03820384 0.025019860 0.049758479 0.033173481 0.033417034
#>      B 0.03160071 0.50000000 0.058205633 0.017813403 0.069379481 0.126121892
#>      C 0.03133429 0.08812724 0.500000000 0.020847662 0.080917264 0.115870617
#>      D 0.09640971 0.04172638 0.032253442 0.500000000 0.039197458 0.038920032
#>      E 0.03043049 0.07694123 0.059268524 0.018557613 0.500000000 0.142042625
#>      F 0.02584792 0.11793923 0.071564226 0.015537357 0.119772862 0.500000000
#>      G 0.03399444 0.02462628 0.012922426 0.012048663 0.018940936 0.019717946
#>      H 0.01575832 0.01283943 0.006520992 0.005817667 0.009055353 0.009798520
#>      I 0.01571555 0.01210172 0.006222032 0.005677597 0.008715936 0.009348947
#>      J 0.07365082 0.04556796 0.033290040 0.045145471 0.052349130 0.045802235
#>       destination
#> origin          G          H          I           J
#>      A 0.09074166 0.08544965 0.08868124 0.055554760
#>      B 0.05437352 0.05758854 0.05648575 0.028431066
#>      C 0.04319941 0.04428422 0.04397130 0.031447997
#>      D 0.06231482 0.06112275 0.06207552 0.065979887
#>      E 0.04637865 0.04504262 0.04511636 0.036221887
#>      F 0.04071159 0.04109779 0.04080585 0.026723172
#>      G 0.50000000 0.15855492 0.19861538 0.020578998
#>      H 0.07805095 0.50000000 0.35315228 0.009006485
#>      I 0.09395277 0.33935970 0.50000000 0.008905744
#>      J 0.07282480 0.06474586 0.06662368 0.500000000

# Simulate point estimates for some mean parameter values
sim_mobility(D, N,
             theta = 3,
             omega_1 = 2,
             omega_2 = 0.7,
             gamma = 1.5,
             tau=rbeta(length(N), 2, 2))

#>       destination
#> origin           A           B            C            D           E
#>      A 0.745898500 0.015067824 0.0079538195 0.0223145468 0.012285977
#>      B 0.028330639 0.344141822 0.0708940692 0.0120066459 0.093343406
#>      C 0.025049254 0.118747135 0.4096032114 0.0136126102 0.105281688
#>      D 0.146626260 0.041960395 0.0284018112 0.3974071099 0.038498471
#>      E 0.031604638 0.127708245 0.0859953355 0.0150716538 0.206076789
#>      F 0.007541249 0.073872921 0.0347776538 0.0035193385 0.076184856
#>      G 0.020439998 0.012666601 0.0047955277 0.0043188447 0.008609881
#>      H 0.007092263 0.005242417 0.0018899126 0.0015930652 0.003128989
#>      I 0.003328027 0.002260246 0.0008299312 0.0007236517 0.001392161
#>      J 0.074778468 0.036575687 0.0227474651 0.0359353455 0.045383756
#>       destination
#> origin           F          G          H          I           J
#>      A 0.012276245 0.05493619 0.04993371 0.05264658 0.026686620
#>      B 0.226106659 0.06400937 0.06939791 0.06722717 0.024542304
#>      C 0.178295998 0.04059132 0.04190538 0.04134704 0.025566371
#>      D 0.037644973 0.07627267 0.07369978 0.07522056 0.084267969
#>      E 0.319030365 0.05952726 0.05667012 0.05665182 0.041663770
#>      F 0.751227021 0.01492222 0.01505442 0.01485300 0.008047325
#>      G 0.009038120 0.43872588 0.20500755 0.28662478 0.009772819
#>      H 0.003480785 0.07825976 0.14908544 0.74711660 0.003110767
#>      I 0.001528456 0.04869766 0.33251719 0.60728151 0.001441160
#>      J 0.036707668 0.07360046 0.06137047 0.06388201 0.549018676

# Simulate multiple stochastic realizations of a mobility matrix from
# fitted mobility model parameters
mod <- summarize_mobility(
  fit_mobility(M, D, N)
)

#> Estimating probability of travel outside origin...

#> These missing locations will inherit population mean:

#> C D F H

#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 6
#>    Unobserved stochastic nodes: 9
#>    Total graph size: 26
#> 
#> Initializing model
#> 
#> NOTE: Stopping adaptation
#> 
#> 

#> Complete.

#> Estimating travel among destinations with gravity model...

#> ::Fitting gravity model for 10 origins and 10 destinations::

#> Using uniformative priors

#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 74
#>    Unobserved stochastic nodes: 30
#>    Total graph size: 1166
#> 
#> Initializing model
#> 
#> NOTE: Stopping adaptation
#> 
#> 

#> Complete.

mod_sim <- sim_mobility(D,
             N,
             n=3,
             mod=mod,
             counts=TRUE)